added registry and roles

+++ src/main/java/com/onserv/overlord/dto/UserDto.java
@@ -19,6 +19,6 @@ public class UserDto {
 
     private Boolean is_active;
 
-    private Integer role_id;
+    private Long role_id;
 
 }
